Ensure all occurrences are replaced in path parameter sanitization
Javascript replace only acts on the first found match if the pattern is
a string. Change the patterns to regular expressions and use the global
flag to ensure all occurrences are replaced.

This also updates the test to include some checks around the three
special characters we un-escape in this code.